fix: jwt bug, rate limiting bug, CORS

+++ src/main/java/com/devilish/planwise/dto/goal/GoalResponse.java
@@ -34,7 +34,7 @@ public class GoalResponse {
             goal.getStatus(),
             goal.getCreatedAt(),
             goal.getActive(),
-            goal.getProgressPercentage()
+            BigDecimal.valueOf(goal.getProgressPercentage())
         );
     }
 }
